requirements.txt
asgiref==3.7.2
boto3==1.34.23
botocore==1.34.23
Django==4.2.9
django-storages==1.14.2
djangorestframework==3.14.0
drf-extra-fields==3.4.1
django-phonenumber-field==7.3.0
jmespath==1.0.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
phonenumbers==8.13.32
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
google-auth==2.28.1
PyJWT==2.8.0
requests==2.31.0
django-treebeard==4.7.1